+++ internal/connector/common/common_utils.go
@@ -7,13 +7,23 @@ import (
 	"text/template"
 
 	"github.com/pingidentity/pingctl/internal/connector"
-	"github.com/rs/zerolog"
+	"github.com/pingidentity/pingctl/internal/logger"
 )
 
-func WriteFiles(exportableResources []connector.ExportableResource, l zerolog.Logger, format, outputDir string, overwriteExport bool) error {
+func WriteFiles(exportableResources []connector.ExportableResource, format, outputDir string, service string, overwriteExport bool) error {
+	l := logger.Get()
+
 	hclImportBlockTemplate, err := template.New("HCLImportBlock").Parse(connector.HCLImportBlockTemplate)
 	if err != nil {
 		return fmt.Errorf("failed to parse HCL import block template. err: %s", err.Error())
+	}
+
+	// Make subdirectory for exported service
+	if len(exportableResources) > 0 {
+		err := os.MkdirAll(filepath.Join(outputDir, service), os.ModePerm)
+		if err != nil {
+			return fmt.Errorf("failed to create %s export subfolder. err: %s", service, err.Error())
+		}
 	}
 
 	for _, exportableResource := range exportableResources {
@@ -31,7 +41,7 @@ func WriteFiles(exportableResources []connector.ExportableResource, l zerolog.Lo
 		l.Debug().Msgf("Generating import file for %s resource...", exportableResource.ResourceType())
 
 		outputFileName := fmt.Sprintf("%s.tf", exportableResource.ResourceType())
-		outputFilePath := filepath.Join(outputDir, filepath.Base(outputFileName))
+		outputFilePath := filepath.Join(outputDir, service, filepath.Base(outputFileName))
 
 		// Check to see if outputFile already exists.
 		// If so, default behavior is to exit and not overwrite.
