+++ internal/connector/pingone/platform/resources/pingone_agreement_localization_enable.go
