+++ internal/connector/pingone/platform/resources/pingone_branding_theme_default.go
