+++ internal/connector/pingone/platform/resources/pingone_gateway.go
