+++ internal/connector/pingone/platform/resources/pingone_gateway_credential.go
