+++ internal/connector/pingone/platform/resources/pingone_gateway_role_assignment.go
