+++ internal/connector/pingone/platform/resources/pingone_notification_settings_email.go
