+++ internal/connector/pingone/platform/resources/pingone_notification_settings_template_content.go
