+++ internal/connector/pingone/platform/resources/pingone_trusted_email_domain.go
