+++ internal/connector/pingone/sso/resources/pingone_resource_scope_openid.go
